Remove redundant and impossible permission sets
1. Some permission sets require displaying menu item in general settings
page. This however can only be rendered if the permission is to manage
the setting so configuration display setting cannot remain unless we
want to change how spree renders the sub menu.

2. Similarly Stock Management can be rolled into one permission. There
is no single view where user can see the stock movement but not have
manage permission. We would have to add another view to provide this
information to display only permission holder.